requirements.txt
numpy
matplotlib
numpy-stl